extracted folder list and item to separate components
